+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Release History
 
-## 1.51.0 (TBD)
+## 1.51.0 (2026-05-13)
 
 ### Snowpark Python API Updates
-
-#### New Features
 
 #### Bug Fixes